+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [[extension] Array#add (#13)](https://github.com/ngarbezza/oow/issues/13)
 - [[new object] DayOfMonth with essential behavior (#14)](https://github.com/ngarbezza/oow/issues/14)
 - [[extension] Function#value(...args)](https://github.com/ngarbezza/oow/issues/25)
+- [[extension] String#reverse()](https://github.com/ngarbezza/oow/issues/18)
 
 ## [1.8.0] - 2019-10-10